measurements back to float

+++ Entities/Measurement.cs
@@ -4,9 +4,9 @@ public class Measurement
 {
     public long Id { get; set; }
     public long GreenhouseId { get; set; }
-    public double Temperature { get; set; }
-    public double Humidity { get; set; }
-    public double Co2 { get; set; }
+    public float Temperature { get; set; }
+    public float Humidity { get; set; }
+    public float Co2 { get; set; }
     public int Light { get; set; }
     public DateTime Timestamp { get; set; }
 
@@ -14,7 +14,7 @@ public class Measurement
     {
     }
 
-    public Measurement(double temperature, double humidity, double co2, int light)
+    public Measurement(float temperature, float humidity, float co2, int light)
     {
         Temperature = temperature;
         Humidity = humidity;
